Distribute fees from current artifact
