+++ modules/core/src/main/scala/org/tessellation/infrastructure/snapshot/GlobalSnapshotConsensus.scala
@@ -5,16 +5,19 @@ import java.security.KeyPair
 import cats.effect.kernel.Async
 import cats.effect.std.{Random, Supervisor}
 
-import org.tessellation.domain.rewards.EpochRewards
 import org.tessellation.kryo.KryoSerializer
 import org.tessellation.schema.balance.Amount
+import org.tessellation.schema.block.DAGBlock
 import org.tessellation.schema.peer.PeerId
+import org.tessellation.schema.transaction.DAGTransaction
+import org.tessellation.schema.{GlobalIncrementalSnapshot, GlobalSnapshotStateProof}
 import org.tessellation.sdk.config.AppEnvironment
 import org.tessellation.sdk.config.types.SnapshotConfig
 import org.tessellation.sdk.domain.cluster.services.Session
 import org.tessellation.sdk.domain.cluster.storage.ClusterStorage
 import org.tessellation.sdk.domain.gossip.Gossip
 import org.tessellation.sdk.domain.node.NodeStorage
+import org.tessellation.sdk.domain.rewards.Rewards
 import org.tessellation.sdk.domain.snapshot.storage.SnapshotStorage
 import org.tessellation.sdk.infrastructure.consensus.Consensus
 import org.tessellation.sdk.infrastructure.metrics.Metrics
@@ -40,7 +43,7 @@ object GlobalSnapshotConsensus {
     environment: AppEnvironment,
     client: Client[F],
     session: Session[F],
-    rewards: EpochRewards[F],
+    rewards: Rewards[F, DAGTransaction, DAGBlock, GlobalSnapshotStateProof, GlobalIncrementalSnapshot],
   ): F[Consensus[F, GlobalSnapshotEvent, GlobalSnapshotKey, GlobalSnapshotArtifact, GlobalSnapshotContext]] =
     Consensus.make[F, GlobalSnapshotEvent, GlobalSnapshotKey, GlobalSnapshotArtifact, GlobalSnapshotContext](
       GlobalSnapshotConsensusFunctions.make[F](
